MNT Delete wrapper function optics around OPTICS (#13271)
* Delete wrapper function optics around OPTICS

optics () is a simple  wrapper around the OPTICS class which instantiates the
class and calls fit. This wrapper exists for DBSCAN or K-Means but not for all
clustering methods. For code homogeneity, I am deleting this function.

* delete optics wrapper from cluster module list

Previous commit forgot to delete the optics wrapper from the cluster module
list. So the test testing the presence of modules failed.